requirements.txt
cohere
google.cloud